requirements.txt
GitPython==3.1.31
PyYAML==6.0
rich==13.4.2
